print token in terminal
