--- demo_server/demo_server_go/readme.md
@@ -1,10 +1,0 @@
-This demo includes a demo html frontend and a server
-fill in your client_ID and client_secret before test
-
-1. set oauth function on github, 
-   homepage url: http://localhost:3000/
-   authorization callback url: http://localhost:3000/login
-
-2. run "go server.go" in the /demo_server/demo_server_go folder
-3. open website "http://localhost:3000/" and click "Github log in"
-4. you can see token in the terminal
